move ocicni from vendors to pkg/
