+++ pkg/ocicni/types.go
@@ -1,13 +1,19 @@
 package ocicni
 
 const (
+	// DefaultInterfaceName is the string to be used for the interface name inside the net namespace
 	DefaultInterfaceName = "eth0"
-	CNIPluginName        = "cni"
-	DefaultNetDir        = "/etc/cni/net.d"
-	DefaultCNIDir        = "/opt/cni/bin"
+	// CNIPluginName is the default name of the plugin
+	CNIPluginName = "cni"
+	// DefaultNetDir is the place to look for CNI Network
+	DefaultNetDir = "/etc/cni/net.d"
+	// DefaultCNIDir is the place to look for cni config files
+	DefaultCNIDir = "/opt/cni/bin"
+	// VendorCNIDirTemplate is the template for looking up vendor specific cni config/executable files
 	VendorCNIDirTemplate = "%s/opt/%s/bin"
 )
 
+// CNIPlugin is the interface that needs to be implemented by a plugin
 type CNIPlugin interface {
 	// Name returns the plugin's name. This will be used when searching
 	// for a plugin by name, e.g.
